backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.29
aiosqlite>=0.20.0
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
openai>=1.14.0
python-docx>=1.1.0
beautifulsoup4>=4.12.3
reportlab>=4.1.0
